Created mock function to emulate key input presses.
